requirements.txt
streamlit
Pillow
pandas
altair
google-generativeai